Blocking mutexes
Update mutex implementations to make blocking
syscalls, rather than polling. This should
ensure fairer scheduling.

On Windows, we now use named mutexes.

On Linux and MacOS, we use blocking flock.
Due to the way signals work in Go we cannot
interrupt a flock syscall, so instead we
allow the syscall to complete and then release
the lock if nothing is waiting for it. To
avoid an unbounded growth of goroutines
making flock syscalls, we ensure that only
one goroutine is making a call at a time.

For compatibility with older implementations,
the Linux implementation also acquires an
abstract domain socket. On Windows, we do
the same with a named semaphore.