Test acquire with sudo envars
The following commit introduces the ability to mock out the sudo envars
so that we can test the acquire as sudo path through the code.

We introduce the mocking framework here so that we can improve the
dependency testing of the acquire and also eventually improve the
quality of the tests using expectations.